+++ class/HPGrowingTextView.m
@@ -180,7 +180,7 @@
 - (void)textViewDidChange:(UITextView *)textView
 {
 	//size of content, so we can set the frame of self
-	NSInteger newSizeH = internalTextView.contentSize.height;
+	CGFloat newSizeH = internalTextView.contentSize.height;
 	if(newSizeH < minHeight || !internalTextView.hasText) newSizeH = minHeight; //not smalles than minHeight
     
 	if (internalTextView.frame.size.height != newSizeH)
